move atomized to new folder
