+++ fedot/core/operations/atomized_model/atomized_model.py
